+++ packages/react-components/src/components/RangeDatePickerV2/index.ts
@@ -1,0 +1,1 @@
+export { RangeDatePickerV2 } from './RangeDatePickerV2';
